갤러리 이미지 삭제

+++ lib/screen/look_photo_screen.dart
@@ -4,9 +4,9 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 
 class LookPhotoScreen extends StatelessWidget {
-  const LookPhotoScreen({required this.photo_info});
+  const LookPhotoScreen({required this.photo_info, required this.myUid});
   final QueryDocumentSnapshot photo_info;
-
+  final String myUid;
   @override
   Widget build(BuildContext context) {
     var galleryService = GalleryService();
@@ -45,13 +45,14 @@ class LookPhotoScreen extends StatelessWidget {
             onPressed: () {},
             icon: Icon(Icons.download),
           ),
-          IconButton(
-            onPressed: () {
-              galleryService.deleteImg(photo_info.id);
-              Navigator.pop(context);
-            },
-            icon: Icon(Icons.delete),
-          )
+          if (photo_info["uid"] == myUid) //내가 올린 이미지인 경우에만 삭제가능
+            IconButton(
+              onPressed: () {
+                galleryService.deleteImg(photo_info.id);
+                Navigator.pop(context);
+              },
+              icon: Icon(Icons.delete),
+            )
         ],
       ),
       body: PageView.builder(
